A better fix for the JSXGraph imageview issue.
There were two connected problems.  First, I was not checking the
correct container div size in the "drawPromise" for the dialog version
of the graph. I was always checking the div for the main JSXGraph image
in the page.  Although, that doesn't entirely matter since the real
problem was that the imageview JSXGraph div did not have its size set on
initialization (which I do in all of the other places where I do this
sort of thing with JSXGraph).